+++ src/typings/global.d.ts
@@ -23,6 +23,11 @@ declare module '*.jpg' {
     export default url;
 }
 
+declare module '*.jpeg' {
+    const url: string;
+    export default url;
+}
+
 declare module '*.png' {
     const url: string;
     export default url;